bunch of verification identity and implementation categories

+++ resources/exchange.py
@@ -2,10 +2,9 @@ from flask_restful import Resource
 from flask import request
 from schemas.exchange import ExchangeSchema
 from schemas.numberlimit import NumberLimitSchema
-from flask_jwt_extended import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt_identity
 from models.exchange import ExchangeModel
 from libs.strings import gettext
-from libs.decorator import verify_permission
 
 exchange_schema = ExchangeSchema()
 exchange_list_schema = ExchangeSchema(many=True)
@@ -24,7 +23,9 @@ class Exchange(Resource):
     @classmethod
     @jwt_required
     def post(cls, id: int):
-        exchange = exchange_schema.load(request.get_json())
+        exchange = exchange_schema.load(request.get_json(), instance=ExchangeModel())
+        if exchange.owner != get_jwt_identity():
+            return {"message": gettext("not_allow")}, 500
         try:
             exchange.save_to_db()
         except:
@@ -33,9 +34,10 @@ class Exchange(Resource):
 
     @classmethod
     @jwt_required
-    # @verify_permission
     def delete(cls, id: int):
         exchange = ExchangeModel.find_by_id(id)
+        if exchange.owner != get_jwt_identity():
+            return {"message": gettext("not_allow")}, 500
         if exchange:
             exchange.delete_from_db()
             return exchange_schema.dump(exchange)
@@ -47,6 +49,8 @@ class Exchange(Resource):
             request.get_json(), instance=ExchangeModel()
         )
         exchange = ExchangeModel.find_by_id(id)
+        if exchange.owner != get_jwt_identity():
+            return {"message": gettext("not_allow")}, 500
         if exchange:
             exchange.name = exchange_data.name
             exchange.description = exchange_data.description
